reformat code; add check on missing file in statistics function

+++ utils/oatest.cpp
@@ -150,7 +150,7 @@ Eigen::MatrixXd array2xfeigen3 ( const array_link &al )
 		}
 	}
 
-	//	mymatrix.block(0,0, n, k+1).array() -= 1; 
+	//	mymatrix.block(0,0, n, k+1).array() -= 1;
 
 	// init interactions
 	ww=k+1;
@@ -275,43 +275,43 @@ int main ( int argc, char* argv[] )
 	{
 		array_link al = exampleArray(23, 1);
 		{
-		jstructconference_t js(al.n_rows, 4);		
-		js.show();		
+		jstructconference_t js(al.n_rows, 4);
+		js.show();
 		js.showdata();
 	}
-		jstructconference_t js(al, 4);		
-		js.show();		
+		jstructconference_t js(al, 4);
+		js.show();
 		js.showdata();
 
-		
-		printf("size %d\n", js.values.size() );
-		
+
+		printf("size %d\n", (int)js.values.size() );
+
 		return 0;
 }
 
 	array_link al = exampleArray(2);
 	al.showarray();
 		Eigen::MatrixXd m1 = arraylink2eigen(array2xf ( al));
-	//std::cout << (m1) << std::endl; exit(0);		
-		
+	//std::cout << (m1) << std::endl; exit(0);
+
 	Eigen::MatrixXd m2 = arraylink2eigen( array2xf2 ( al) );
 
 	std::cout << (m1-m2) << std::endl;
-	
+
 	for(int i=0; i<1000000; i++) {
 	//Eigen::MatrixXd m = array2xfeigen ( al);
 	//Eigen::MatrixXd m = arraylink2eigen(array2xf ( al) );
 	Eigen::MatrixXd m = arraylink2eigen(array2xf2( al) );
 	}
-	
+
 	exit(0);
-	
+
 	if (1) {
 		arraylist_t ll = readarrayfile ( "dummy-24-4.oa" );
 		array_link al=ll[0];
 		al.show();
 
-int N=al.n_rows;		
+int N=al.n_rows;
 				conference_t ctype(N, N);
 
 
@@ -319,19 +319,19 @@ int N=al.n_rows;
 	array_link al3 = ctype.create_root_three();
 
 	//arraylist_t lst; lst.push_back(al3); writearrayfile("test.oa", lst, arrayfile::ABINARY);
-	
+
 	int extcol=6;
 	cperm_list ee= generateConferenceExtensions ( al2, ctype, extcol, 0, 0, 1 );
 	printf("generated %d\n", (int)ee.size() );
 
 	 ee= generateConferenceExtensions ( al3, ctype, extcol, 0, 0, 1 );
 	printf("generated %d\n", (int)ee.size() );
-		
+
 	DconferenceFilter filter(al, 0,1,0);
-	
+
 	filter.filterList(ee,1);
-	
-	exit(0);	
+
+	exit(0);
 	}
 
 	if ( 1 ) {
@@ -340,20 +340,20 @@ int N=al.n_rows;
 		//arraylist_t ll = readarrayfile ( "/home/eendebakpt/oatmp/conf/dconferencej1j3-36-8.oa" );
 		//arraylist_t ll = readarrayfile ( "/home/eendebakpt/oatmp/conf/dconferencej1j3-28-4.oa" );
 		arraylist_t ll = readarrayfile ( "/home/eendebakpt/oatmp/conf/dconferencej1j3-32-8.oa" );
-		
-		//ll[52]=ll[51]; 
-		//ll[51]=ll[52]; 
-		
+
+		//ll[52]=ll[51];
+		//ll[51]=ll[52];
+
 		//ll[0].Fvalues(4);
 		//ll[0].Fvalues(5);
-		
+
 		array_link alx = ll[0];
 		alx.showproperties();
 		int N = alx.n_rows;
 
 		int fi=51;
 		printf("first:\n"); ll[fi].showarray(); printf("next :\n"); ll[fi+1].showarray(); printf("  first diff index: %d\n", ll[fi].firstColumnDifference(ll[fi+1]) );
-		
+
 
 		conference_t ct(N, 2*N);
 		CandidateGeneratorDouble cgenerator(array_link() , ct);
@@ -362,11 +362,11 @@ int N=al.n_rows;
 
 double t0;
 if (0) {
-t0=get_time_ms();		
+t0=get_time_ms();
 
 				for(size_t i=0; i<ll.size(); i++) {
 			const array_link &al = ll[i];
-			std::vector<cperm> cc1 = generateDoubleConferenceExtensionsInflate (al, ct, 1, 1, 1);		
+			std::vector<cperm> cc1 = generateDoubleConferenceExtensionsInflate (al, ct, 1, 1, 1);
 		}
 				printf ( "   dtt %.1f [ms]\n", 1e3* ( get_time_ms()-t0 ) );
 }
@@ -384,18 +384,18 @@ t0=get_time_ms();
 				printfd("   _________________________\n");
 			}
 			//cperm tmp= cc1[0]; printf("size cc1: %d\n", (int)tmp.size() );
-			
+
 			cgenerator.verbose=2;
 			std::vector<cperm> cc2 = cgenerator.generateDoubleConfCandidates(al);
 			//printf("size cc2: %d\n", cc2[0].size());
-			
-			
-			
+
+
+
 			int nc2=cc2.size();
-			
-		printf("%d: number of candidates: %d/%d\n", (int)i, nc1, nc2);	
+
+		printf("%d: number of candidates: %d/%d\n", (int)i, nc1, nc2);
 		cgenerator.showCandidates();
-		
+
 		if(i==52 && 0) {
 			printf("cc1: \n");
 			showCandidates(cc1);
@@ -612,4 +612,4 @@ t0=get_time_ms();
 
 }
 
-// kate: indent-mode cstyle; indent-width 4; replace-tabs off; tab-width 4; 
+// kate: indent-mode cstyle; indent-width 4; replace-tabs off; tab-width 4;
